ai-python/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
argcomplete==3.6.2
blinker==1.9.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cerebras_cloud_sdk
click==8.3.0
colorama==0.4.6
cryptography==46.0.1
distro==1.9.0
filelock==3.19.1
Flask==2.2.5
flask-cors==6.0.1
fsspec==2025.9.0
greenlet==3.2.4
groq==0.31.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.76
langchain-text-splitters==0.3.11
langsmith==0.4.30
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
orjson==3.11.3
packaging==25.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
pipx==1.7.1
platformdirs==4.4.0
psycopg2-binary==2.9.10
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pypdfium2==4.30.0
pytesseract==0.3.13
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.1
setuptools==80.9.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
transformers==4.56.2
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
userpath==1.9.2
Werkzeug==3.1.3
zstandard==0.25.0
docx2txt
pandas
python-pptx
